requirements.txt
opencv-python
face-recognition
numpy
psycopg2-binary
scipy
fastapi
uvicorn
